Refactor larpix.py into submodules
